refactor(config): change Module specification to easy to understand

+++ util.h
@@ -27,7 +27,7 @@ typedef struct _list_head {
 #define list_for_each(head, pos) \
 for (pos = (head)->next; pos != (head); pos = pos->next)
 #define list_entry(ptr, type, member) \
-((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+((type *)((char *)(ptr)-(intptr_t)(&((type *)0)->member)))
 #define list_empty(head) ((head)->next == (head))
 #define list_count(head, num) \
 for (list_head *pos = (head)->next; pos != (head); pos = pos->next) num++
